added monitor options for notifications
