+++ modules/osd/index.ts
@@ -1,9 +1,9 @@
-import { OSDAnchor } from "lib/types/options";
 import options from "options";
 import brightness from "services/Brightness"
 import { OSDLabel } from "./label/index";
 import { OSDBar } from "./bar/index";
 import { OSDIcon } from "./icon/index";
+import { getPosition } from "lib/utils";
 const hyprland = await Service.import("hyprland");
 const audio = await Service.import("audio")
 
@@ -38,20 +38,6 @@ const handleReveal = (self: any, property: string) => {
     })
 }
 
-const getPosition = (pos: OSDAnchor): ("top" | "bottom" | "left" | "right")[] => {
-    const positionMap: { [key: string]: ("top" | "bottom" | "left" | "right")[] } = {
-        "top": ["top"],
-        "top right": ["top", "right"],
-        "top left": ["top", "left"],
-        "bottom": ["bottom"],
-        "bottom right": ["bottom", "right"],
-        "bottom left": ["bottom", "left"],
-        "right": ["right"],
-        "left": ["left"],
-    };
-
-    return positionMap[pos];
-}
 const renderOSD = () => {
